move handleOpenImagePopup to index.js
