+++ src/components/index.js
@@ -1,8 +1,11 @@
 import '../pages/index.css'; // для webpack (и убрать подключение css в index.html)
 
-import {partialCreateCard} from './card.js';
+import {createPartialCreateCard, deleteCard, handleLikeClick} from './card.js';
 import {openPopup, closePopup} from './modal.js';
 import {initialCards} from './cards.js';
+
+// Задаем постоянные (неизменяемые) параметры для partial функции создания карточки
+const partialCreateCard = createPartialCreateCard(deleteCard, handleOpenImagePopup, handleLikeClick);
 
 const placeList = document.querySelector('.places__list');
 
@@ -16,7 +19,7 @@ const addCartButton = document.querySelector('.profile__add-button');
 // Модальные окна
 const editProfilePopup = document.querySelector('.popup_type_edit');
 const addCardPopup = document.querySelector('.popup_type_new-card');
-//const openImagePopup = document.querySelector('.popup_type_image');
+const openImagePopup = document.querySelector('.popup_type_image');
 
 // Переменные профиля
 const profileTitle = document.querySelector('.profile__title');
@@ -55,6 +58,15 @@ const handleCardFormSubmit = async (event) => {
     placeList.prepend(partialCreateCard(card));
     closePopup(document.querySelector('.popup_type_new-card'));
     cardForm.reset() // очищаем форму
+}
+
+// Обрабатываем открытие попапа с картинкой (передается в функцию createCard)
+function handleOpenImagePopup(card) {
+    const imagePopup = document.querySelector('.popup__image');
+    const titlePopup = document.querySelector('.popup__caption');
+    imagePopup.src = card.link;
+    imagePopup.alt = titlePopup.textContent = card.name;
+    openPopup(openImagePopup);
 }
